requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==10.4.0
environs[django]==9.3.2
django-phonenumber-field==6.1.0
phonenumbers==8.13.47
djangorestframework==3.12.4
